Product tab on home screen is getting data
Product part on home screen is working
