+++ app/src/main/java/com/cegep/sportify/Adapter/ProductAdapter.java
@@ -16,25 +16,25 @@ import com.cegep.sportify.model.Product;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
 
 public class ProductAdapter extends RecyclerView.Adapter<ProductViewHolder> {
 
     private final Context context;
-    private final Product productobj;
 
-    private final List<Product> products = new ArrayList<>();
+    private final List<Product> products;
 
-    public ProductAdapter(Context context, Product clickedProduct) {
+    public ProductAdapter(Context context, List<Product> products) {
         this.context = context;
-        this.productobj = clickedProduct;
+        this.products = products;
     }
 
     @Override
     public ProductViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
         LayoutInflater inflater = LayoutInflater.from(context);
         View view = inflater.inflate(R.layout.item_product, parent, false);
-        return new ProductViewHolder(view,productobj);
+        return new ProductViewHolder(view);
     }
 
     @Override
@@ -45,5 +45,11 @@ public class ProductAdapter extends RecyclerView.Adapter<ProductViewHolder> {
     @Override
     public int getItemCount() {
         return products.size();
+    }
+
+    public void update(Collection<Product> products) {
+        this.products.clear();
+        this.products.addAll(products);
+        notifyDataSetChanged();
     }
 }
