+++ app/src/main/res/layout/activity_main.xml
@@ -32,11 +32,23 @@
 
             <androidx.fragment.app.FragmentContainerView
                 android:id="@+id/container"
-                android:name="com.cegep.sportify.DashboardFragment"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent" />
 
         </com.google.android.material.appbar.AppBarLayout>
+
+        <com.google.android.material.bottomnavigation.BottomNavigationView
+            android:id="@+id/bottom_navigation"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_gravity="bottom"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintHorizontal_bias="0.0"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="@id/container"
+            app:layout_constraintVertical_bias="1.0"
+            app:menu="@menu/bottom_menu_dashboard" />
 
     </androidx.coordinatorlayout.widget.CoordinatorLayout>
 
@@ -48,5 +60,4 @@
         android:fitsSystemWindows="true"
         app:headerLayout="@layout/header_layout"
         app:menu="@menu/drawer_menu" />
-
 </androidx.drawerlayout.widget.DrawerLayout>
