+++ app/src/main/res/layout/fragment_products_list.xml
@@ -3,6 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:background="@color/white"
     tools:context=".ProductsListFragment" >
 
     <androidx.recyclerview.widget.RecyclerView
