+++ app/src/main/res/layout/item_product.xml
@@ -9,6 +9,10 @@
     android:layout_marginTop="@dimen/size_l"
     app:elevation="@dimen/size_l">
 
+    <ScrollView
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        >
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
@@ -91,5 +95,7 @@
             tools:text="$25.99" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
+
+    </ScrollView>
 
 </com.google.android.material.card.MaterialCardView>
